Facerecognition/requirements.txt
streamlit
opencv-python
pillow
tensorflow
torch
torchvision
scikit-learn
keras
numpy
pandas
sqlalchemy
pymysql
gitpython
matplotlib
seaborn
requests
joblib
face_recognition
glob